voice-agents/knowledge_base_rag/pyproject.toml
[project]
name = "atoms-knowledge-base-rag"
version = "0.1.0"
description = "Knowledge Base setup and RAG-powered voice agent"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "smallestai>=4.2.0",
    "python-dotenv",
    "loguru",
]
